% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulletchartr_data.r
\docType{data}
\name{bc_ex}
\alias{bc_ex}
\title{example dataframe for regular scale bullet chart}
\format{This data set contains indicator values and target data
used in the examples in README for the regular scale bullet charts.

The variables are as follows:

 \itemize{
  \item variable. name of the indicator
  \item info. extra info about indicator
  \item target. target value of indicator
  \item current. current value of indicator
  \item low. low value of indicator
  \item medium. medium value of indicator
  \item high. high value of indicator

 }}
\usage{
data(bc_ex)
}
\description{
example dataframe for regular scale bullet chart
}
\keyword{datasets}
