% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulletchartr_data.r
\docType{data}
\name{bc_ex_TC}
\alias{bc_ex_TC}
\title{example dataframe for time-comparison scale bullet chart}
\format{This data set contains indicator values and target data
used in the examples in README for the time-comparison scale bullet charts.

The variables are as follows:

 \itemize{
  \item indicator_name. name of the indicator
  \item actual. present value of indicator
  \item actual_lastweek. value of indicator in previous week
  \item actual_lastyear. value of indicator on same day of last year
  \item target. target value of indicator

 }}
\usage{
data(bc_ex_TC)
}
\description{
example dataframe for time-comparison scale bullet chart
}
\keyword{datasets}
