% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet_chart_functions.r
\name{bullet_chart}
\alias{bullet_chart}
\title{bullet_chart}
\usage{
bullet_chart(file_name, sheet_name = "Sheet1", for_year = year(Sys.Date()),
  FY = TRUE, project_start_date)
}
\arguments{
\item{sheet_name}{specify which sheet in Excel file, Default: "Sheet1"}

\item{for_year}{specify the year in which the report is being made, Default: year(Sys.Date())}

\item{FY}{current fiscal year, Default: TRUE}

\item{project_start_date}{specify start date of the project as %Y/%D/%M string}

\item{fileName}{path of Excel file}
}
\value{
bullet chart with symbols for "last week" and "last year"
}
\description{
creates bullet chart with symbols
}
\details{
DETAILS...
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{geom_col}}
}
