% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet_chart_functions.r
\name{bullet_chart3}
\alias{bullet_chart3}
\title{bullet_chart3}
\usage{
bullet_chart3(file_name, sheet_name = "Sheet1", for_year = year(Sys.Date()),
  FY = TRUE, project_start_date)
}
\arguments{
\item{file_name}{PARAM_DESCRIPTION}

\item{sheet_name}{PARAM_DESCRIPTION, Default: 'Sheet1'}

\item{for_year}{PARAM_DESCRIPTION, Default: year(Sys.Date())}

\item{FY}{PARAM_DESCRIPTION, Default: TRUE}

\item{project_start_date}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{ggplot}}
}
