% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet_chart_functions.r
\name{bullet_chart4}
\alias{bullet_chart4}
\title{bullet_chart4}
\usage{
bullet_chart4(file_name, sheet_name = "Sheet1", for_year = year(Sys.Date()),
  FY = TRUE, project_start_date)
}
\arguments{
\item{file_name}{path of Excel file}

\item{sheet_name}{specify which sheet in Excel file, Default: "Sheet1"}

\item{for_year}{specify the year in which the report is being made, Default: year(Sys.Date())}

\item{FY}{fiscal year or calendar year? Default: TRUE}

\item{project_start_date}{specify start date of the project as \%Y/\%D/\%M format string}
}
\description{
create bullet chart showing last year's value as the target
}
\details{
This version of the bullet chart shows a single colored bar representing the current value
for the indicator along with a black vertical line representing the indicator value at this time
last year. The definition for the vertical line can be changed to your preference (such as a more
general "Target" value), however at the current time you should change the values of "Actual_lastYear"
in the Excel file but not the variable name itself.
}
\examples{
bullet_chart4("data/Indicators_Targets_ext.xlsx", project_start_date = "2016/03/01")
}
\seealso{
\code{\link[ggplot2]{ggplot}}
}
