% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet_chart_functions.r
\name{bullet_chart_symbols}
\alias{bullet_chart_symbols}
\title{bullet_chart_symbols}
\usage{
bullet_chart_symbols(file_name = NULL, sheet_name = "Sheet1",
  dataframe = NULL, indicator_name = "indicator_name", actual = "actual",
  actual_lastweek = "actual_lastweek", actual_lastyear = "actual_lastyear",
  target = "target", for_year = year(Sys.Date()), cal_type = "fis",
  small = FALSE)
}
\arguments{
\item{file_name}{path of Excel file}

\item{sheet_name}{specify which sheet in Excel file, Default: "Sheet1"}

\item{dataframe}{specify R dataframe input}

\item{indicator_name}{specify the name of the column that has your indicator/KPI names}

\item{actual}{specify the name of the column that has the current value of your indicators/KPIs}

\item{actual_lastweek}{specify the name of the column that has the indicator/KPI value from the previous week}

\item{actual_lastyear}{specify the name of the column that has the indicator/KPI value from the previous year}

\item{target}{specify the name of the column that has the target value for the indicator/KPI}

\item{for_year}{specify the year in which the report is being made, Default: year(Sys.Date())}

\item{cal_type}{define what calendar you are using. Options are "fis" for fiscal year starting
October 1st, "cal" for calendar year starting January 1st, or enter your own custom date in the
format "YYYY/MM/DD", Default: fis}

\item{small}{specify whether you want the small version of the plot (TRUE or FALSE), Default: FALSE}
}
\description{
creates bullet chart with symbols
}
\details{
The bar for each Indicator show the progression along the horizontal-axis presenting
the percentage of the yearly target completed. This axis also shows the percent of the year
gone by with the vertical line indicating what exact percentage "Today" is, along this percentage.

Each bar is colored "green" if the Indicator completion is close to or past "Today", "orange" if
it's close, and "red" if it is far off track.

The symbols represent the indicator value for last week (diamond) and last year (circle).
}
\examples{
 bullet_chart_symbols(file_name = system.file("data/Indicators_targets.xlsx",
 package="bulletchartr"))
}
\seealso{
\code{\link[ggplot2]{geom_bar}}, \code{\link[ggplot2]{scale_manual}}
}
