% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet_chart_functions.r
\name{bullet_chart_vline}
\alias{bullet_chart_vline}
\title{bullet_chart_vline}
\usage{
bullet_chart_vline(file_name, sheet_name = "Sheet1",
  for_year = year(Sys.Date()), cal_type = "fis")
}
\arguments{
\item{file_name}{path of Excel file}

\item{sheet_name}{specify which sheet in Excel file, Default: "Sheet1"}

\item{for_year}{specify the year in which the report is being made, Default: year(Sys.Date())}

\item{cal_type}{define what calendar you are using. Options are "fis" for fiscal year starting
October 1st, "cal" for calendar year starting January 1st, or enter your own custom date in the
format "YYYY/MM/DD", Default: fis}
}
\description{
create bullet chart showing last year's value as the target
}
\details{
This version of the bullet chart shows a single colored bar representing the current value
for the indicator along with a black vertical line representing the indicator value at this time
last year. The definition for the vertical line can be changed to your preference (such as a more
general "target" value), however at the current time you should change the values of "actual_lastyear"
in the Excel file but not the variable name itself.
}
\examples{
bullet_chart_vline("data/Indicators_targets_ext.xlsx")
}
\seealso{
\code{\link[ggplot2]{ggplot}}
}
