% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulletchart222.R
\name{bulletchart}
\alias{bulletchart}
\title{bulletchart}
\usage{
bulletchart(file_name = NULL, sheet_name = "Sheet1",
  dataframe = NULL, indicator_name = "variable", info = "info",
  current = "current", low = "low", medium = "medium",
  high = "high", target = "target", remove_no_targets = TRUE,
  legend = TRUE)
}
\arguments{
\item{file_name}{PARAM_DESCRIPTION, Default: NULL}

\item{sheet_name}{PARAM_DESCRIPTION, Default: 'Sheet1'}

\item{dataframe}{PARAM_DESCRIPTION, Default: NULL}

\item{indicator_name}{PARAM_DESCRIPTION, Default: 'variable'}

\item{info}{PARAM_DESCRIPTION, Default: 'info'}

\item{current}{PARAM_DESCRIPTION, Default: 'current'}

\item{low}{PARAM_DESCRIPTION, Default: 'low'}

\item{medium}{PARAM_DESCRIPTION, Default: 'medium'}

\item{high}{PARAM_DESCRIPTION, Default: 'high'}

\item{target}{PARAM_DESCRIPTION, Default: 'target'}

\item{remove_no_targets}{PARAM_DESCRIPTION, Default: TRUE}

\item{legend}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
Stephen Few style bullet chart
}
