% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet_chart_functions.r
\name{extra_field_calculator}
\alias{extra_field_calculator}
\title{extra_field_calculator}
\usage{
extra_field_calculator(fileName, for_year = year(Sys.Date()), FY = TRUE,
  project_start_date = NULL)
}
\arguments{
\item{fileName}{PARAM_DESCRIPTION}

\item{for_year}{PARAM_DESCRIPTION, Default: year(Sys.Date())}

\item{FY}{PARAM_DESCRIPTION, Default: TRUE}

\item{project_start_date}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
