% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_bc.R
\name{field_calculator}
\alias{field_calculator}
\title{Field calculator for regular scale}
\usage{
field_calculator(
  file_name = NULL,
  sheet_name = "Sheet1",
  dataframe = NULL,
  indicator_name = "variable",
  info = "info",
  current = "current",
  low = "low",
  medium = "medium",
  high = "high",
  target = "target",
  remove_no_targets = TRUE
)
}
\arguments{
\item{file_name}{path of Excel file}

\item{sheet_name}{specify which sheet in Excel file}

\item{dataframe}{specify R dataframe input}

\item{indicator_name}{specify the name of the column that has your indicator/KPI names}

\item{info}{extra info for indicators (currency, percentage, symbol, etc.)}

\item{current}{current value of indicator}

\item{low}{column with value for "low"}

\item{medium}{column with value for "medium"}

\item{high}{column with value for "high"}

\item{target}{column with value for the "target"}

\item{remove_no_targets}{Removes indicators without a "target" value specified, default: TRUE}
}
\description{
internal function for calculating the extra fields needed for bullet charts
}
\details{
internal function for calculating the extra fields for the bullet chart
}
