% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_example.R
\name{read_example}
\alias{read_example}
\title{Get path to bulletchartr example files}
\usage{
read_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If `NULL`, the example files will be listed.}
}
\description{
bulletchartr comes bundled with a number of sample files in its `inst/extdata`
directory. This function make them easy to access. Taken from 'readr' package.
}
\examples{
read_example()
read_example("Indicators_Targets.xlsx")
}
