% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulletchart.R
\name{bullet_chart}
\alias{bullet_chart}
\title{'Stephen Few'-style Bullet Chart}
\usage{
bullet_chart(
  file_name = NULL,
  sheet_name = "Sheet1",
  dataframe = NULL,
  indicator_name = "variable",
  info = "info",
  current = "current",
  low = "low",
  medium = "medium",
  high = "high",
  target = "target",
  remove_no_targets = TRUE,
  legend = TRUE
)
}
\arguments{
\item{file_name}{path of Excel file}

\item{sheet_name}{specify which sheet in Excel file, Default: "Sheet1"}

\item{dataframe}{specify R dataframe input}

\item{indicator_name}{specify the name of the column that has your indicator/KPI names,
Default: 'variable'}

\item{info}{extra info for indicators (currency, percentage, symbol, etc.)}

\item{current}{current value of indicator}

\item{low}{column with value for "low"}

\item{medium}{column with value for "medium"}

\item{high}{column with value for "high"}

\item{target}{column with value for the "target"}

\item{remove_no_targets}{Removes indicators without a "target" value specified, default: TRUE}

\item{legend}{Show legend, Default: TRUE}
}
\value{
bullet chart plot(s)
}
\description{
Creates a bullet chart using an indicator's values for the axis scales.
}
\details{
Stephen Few style bullet chart
}
\examples{
data("bc_ex")
bullet_chart(dataframe = bc_ex)
}
