% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet_chart_functions.r
\name{bullet_chart}
\alias{bullet_chart}
\title{bullet_chart}
\usage{
bullet_chart(file_name = NULL, sheet_name = "Sheet1",
  dataframe = NULL, indicator_name = "indicator_name",
  actual = "actual", actual_lastweek = "actual_lastweek",
  actual_lastyear = "actual_lastyear", target = "target",
  for_year = year(Sys.Date()), cal_type = "fis", small = FALSE,
  legend = FALSE, remove_no_targets = FALSE)
}
\arguments{
\item{file_name}{path of Excel file}

\item{sheet_name}{specify which sheet in Excel file, Default: "Sheet1"}

\item{dataframe}{specify R dataframe input}

\item{indicator_name}{specify the name of the column that has your indicator/KPI names}

\item{actual}{specify the name of the column that has the current value of your indicators/KPIs}

\item{actual_lastweek}{specify the name of the column that has the indicator/KPI value from the previous week}

\item{actual_lastyear}{specify the name of the column that has the indicator/KPI value from the previous year}

\item{target}{specify the name of the column that has the target value for the indicator/KPI}

\item{for_year}{specify the year in which the report is being made, Default: year(Sys.Date())}

\item{cal_type}{define what calendar you are using. Options are "fis" for fiscal year starting
October 1st, "cal" for calendar year starting January 1st, or enter your own custom date in the
format "YYYY/MM/DD", Default: fis}

\item{small}{specify whether you want the small version of the plot (TRUE or FALSE), Default: FALSE}

\item{legend}{specify whether you want to show the legend, Default: FALSE}

\item{remove_no_targets}{remove indicators with Targets == NA or 0, Default: FALSE}
}
\description{
create a Stephen Few bullet chart
}
\details{
This version of the bullet chart most closely resembles Stephen Few's design. The single black bar represents
the current value of the indicator while the different hue columns represent last week's value (darker hue) and last year's value (lighter hue).
}
\examples{
data(df)
bullet_chart(dataframe = df)
}
\seealso{
\code{\link[ggplot2]{ggplot}}
}
