% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet_chart_functions.r
\name{bullet_chart}
\alias{bullet_chart}
\title{bullet_chart}
\usage{
bullet_chart(file_name, sheet_name = "Sheet1", for_year = year(Sys.Date()),
  cal_type = "fis")
}
\arguments{
\item{file_name}{path of Excel file}

\item{sheet_name}{specify which sheet in Excel file, Default: "Sheet1"}

\item{for_year}{specify the year in which the report is being made, Default: year(Sys.Date())}

\item{cal_type}{define what calendar you are using. Options are "fis" for fiscal year starting
October 1st, "cal" for calendar year starting January 1st, or enter your own custom date in the
format "YYYY/MM/DD", Default: fis}
}
\description{
create a Stephen Few bullet chart
}
\details{
This version of the bullet chart most closely resembles Stephen Few's design. The single black bar represents
the current value of the indicator while the different hue columns represent last week's value (darker hue) and last year's value (lighter hue).
}
\examples{
bullet_chart3("data/Indicators_targets_ext.xlsx", project_start_date = "2016/03/01")
}
\seealso{
\code{\link[ggplot2]{ggplot}}
}
