% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet_chart_functions.r
\name{bullet_chart2}
\alias{bullet_chart2}
\title{bullet_chart2}
\usage{
bullet_chart2(file_name, sheet_name = "Sheet1", for_year = year(Sys.Date()),
  cal_type = "fis")
}
\arguments{
\item{file_name}{path of Excel file}

\item{cal_type}{define what calendar you are using. Options are "fis" for fiscal year starting
October 1st, "cal" for calendar year starting January 1st, or enter your own custom date in the
format "YYYY/MM/DD", Default: fis}

\item{FY}{fiscal year or calendar year? Default: TRUE}
}
\description{
create bullet chart with bars of varying width
}
\details{
This version conforms more closely with the standard bullet chart design. This function
uses different thicknesses for the bars as the benchmarks for previous time points (last week and last year) to further
accentuate the difference graphically.
}
\examples{
bullet_chart2(file_name = "data/Indicators_targets.xlsx")

}
\seealso{
\code{\link[ggplot2]{geom_bar}}
}
