% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet_chart_functions.r
\name{bullet_chart3}
\alias{bullet_chart3}
\title{bullet_chart3}
\usage{
bullet_chart3(file_name, sheet_name = "Sheet1", for_year = year(Sys.Date()),
  FY = TRUE, project_start_date)
}
\arguments{
\item{file_name}{path of Excel file}

\item{sheet_name}{specify which sheet in Excel file, Default: "Sheet1"}

\item{for_year}{specify the year in which the report is being made, Default: year(Sys.Date())}

\item{FY}{fiscal year or calendar year? Default: TRUE}

\item{project_start_date}{specify start date of the project as \%Y/\%D/\%M format string}
}
\description{
creates bullet chart with symbols
}
\details{
The bar for each Indicator show the progression along the horizontal-axis presenting
the percentage of the yearly target completed. This axis also shows the percent of the year
gone by with the vertical line indicating what exact percentage "Today" is, along this percentage.

Each bar is colored "green" if the Indicator completion is close to or past "Today", "orange" if
it's close, and "red" if it is far off track.

The symbols represent the indicator value for last week (diamond) and last year (circle).
}
\examples{
 bullet_chart(file_name = "data/Indicators_Targets.xlsx", project_start_date = "2016/03/01")
}
\seealso{
\code{\link[ggplot2]{geom_bar}}, \code{\link[ggplot2]{scale_manual}}
}
