% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet_chart_wide.R
\name{bullet_chart_wide}
\alias{bullet_chart_wide}
\title{bullet_chart_wide}
\usage{
bullet_chart_wide(
  file_name = NULL,
  sheet_name = "Sheet1",
  dataframe = NULL,
  indicator_name = "indicator_name",
  actual = "actual",
  actual_lastweek = "actual_lastweek",
  actual_lastyear = "actual_lastyear",
  target = "target",
  for_year = year(Sys.Date()),
  cal_type = "fis",
  chart_type = "static",
  small = FALSE,
  legend = TRUE,
  remove_no_targets = FALSE,
  show_text = FALSE
)
}
\arguments{
\item{file_name}{path of Excel file}

\item{sheet_name}{specify which sheet in Excel file, Default: "Sheet1"}

\item{dataframe}{specify R dataframe input}

\item{indicator_name}{specify the name of the column that has your indicator/KPI names}

\item{actual}{specify the name of the column that has the current value of your indicators/KPIs}

\item{actual_lastweek}{specify the name of the column that has the indicator/KPI value from the previous week}

\item{actual_lastyear}{specify the name of the column that has the indicator/KPI value from the previous year}

\item{target}{specify the name of the column that has the target value for the indicator/KPI}

\item{for_year}{specify the year in which the report is being made, Default: year(Sys.Date())}

\item{cal_type}{define what calendar you are using. Options are "fis" for fiscal year starting
October 1st, "cal" for calendar year starting January 1st, or enter your own custom date in the
format "YYYY/MM/DD", Default: fis}

\item{chart_type}{static of interactive (ggiraph) version}

\item{small}{specify whether you want the small version of the plot (TRUE or FALSE), Default: FALSE}

\item{legend}{specify whether you want to show the legend, Default: TRUE}

\item{remove_no_targets}{remove indicators with Targets == NA or 0, Default: FALSE}

\item{show_text}{Show 'Last Week' & 'Last Year' text, when `small = TRUE` or
`chart_type = "interactive"` then no text will be shown by default.}
}
\description{
create bullet chart with bars of varying width
}
\details{
This version conforms more closely with the standard bullet chart design. This function
uses different thicknesses for the bars as the benchmarks for previous time points (last week and last year) to further
accentuate the difference graphically.
}
\examples{
data("bc_ex_TC")
bullet_chart_symbols(dataframe = bc_ex_TC)
}
\seealso{
\code{\link[ggplot2]{geom_bar}}
}
