% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet_chart_functions.r
\name{extra_field_calculator}
\alias{extra_field_calculator}
\title{extra_field_calculator}
\usage{
extra_field_calculator(file_name = NULL, sheet_name = "Sheet1",
  dataframe = NULL, indicator_name = "indicator_name",
  actual = "actual", actual_lastweek = "actual_lastweek",
  actual_lastyear = "actual_lastyear", target = "target",
  for_year = year(Sys.Date()), cal_type = "fis")
}
\arguments{
\item{file_name}{path of Excel file}

\item{sheet_name}{specify which sheet in Excel file}

\item{dataframe}{specify R dataframe input}

\item{indicator_name}{specify the name of the column that has your indicator/KPI names}

\item{actual}{specify the name of the column that has the current value of your indicators/KPIs}

\item{actual_lastweek}{specify the name of the column that has the indicator/KPI value from the previous week}

\item{actual_lastyear}{specify the name of the column that has the indicator/KPI value from the previous year}

\item{target}{specify the name of the column that has the target value for the indicator/KPI}

\item{for_year}{specify the year in which the report is being made, Default: year(Sys.Date())}

\item{cal_type}{define what calendar you are using. Options are "fis" for fiscal year starting
October 1st, "cal" for calendar year starting January 1st, or enter your own custom date in the
format "YYYY/MM/DD", Default: fis}
}
\description{
internal function for calculating the extra fields needed for bullet charts
}
\details{
internal function for calculating the extra fields for the bullet chart
}
\seealso{
\code{\link[dplyr]{mutate}}
 \code{\link[readxl]{read_excel}}
}
