% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet_chart_functions.r
\name{extra_field_calculator}
\alias{extra_field_calculator}
\title{extra_field_calculator}
\usage{
extra_field_calculator(file_name, sheet_name = "Sheet1",
  for_year = year(Sys.Date()), FY = TRUE, project_start_date)
}
\arguments{
\item{file_name}{path of Excel file}

\item{sheet_name}{specify which sheet in Excel file}

\item{for_year}{specify the year in which the report is being made, Default: year(Sys.Date())}

\item{FY}{fiscal year or calendar year? Default: TRUE}

\item{project_start_date}{specify start date of the project as \%Y/\%D/\%M format string}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
internal function for calculating the extra fields needed for bullet charts
}
\details{
DETAILS
}
\examples{
extra_field_calculator("data/Indicators_Targets.xlsx", for_year = 2018, FY = TRUE,
project_start_date = project_start_date)
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
 \code{\link[readxl]{read_excel}}
}
