% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet_chart_functions.r
\name{extra_field_calculator}
\alias{extra_field_calculator}
\title{extra_field_calculator}
\usage{
extra_field_calculator(fileName, ShetName = "Sheet1",
  for_year = year(Sys.Date()), FY = TRUE, project_start_date = NULL)
}
\arguments{
\item{fileName}{PARAM_DESCRIPTION}

\item{for_year}{PARAM_DESCRIPTION, Default: year(Sys.Date())}

\item{FY}{PARAM_DESCRIPTION, Default: TRUE}

\item{project_start_date}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
extra_field_calculator("data/Indicators_Targets.xlsx", for_year = 2018, FY = TRUE,
project_start_date = project_start_date)
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
 \code{\link[xlsx]{read.xlsx}}
}
