% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal2.R
\name{field_calculator}
\alias{field_calculator}
\title{extra_field_calculator}
\usage{
field_calculator(file_name = NULL, sheet_name = "Sheet1",
  dataframe = NULL, indicator_name = "variable", info = "info",
  current = "current", bad = "bad", good = "good", great = "great",
  target = "target", remove_no_targets = TRUE)
}
\arguments{
\item{file_name}{path of Excel file}

\item{sheet_name}{specify which sheet in Excel file}

\item{dataframe}{specify R dataframe input}

\item{indicator_name}{specify the name of the column that has your indicator/KPI names}

\item{info}{extra info for indicators (currency, percentage, symbol, etc.)}

\item{current}{current value of indicator}

\item{bad}{column with value for "bad"}

\item{good}{column with value for "good"}

\item{great}{column with value for "great", if "target" exceeds "great" then
"great" will be increased to "target" value. If "current" value exceeds "great"
then "current" value will be decreased to "great" value.}

\item{target}{column with value for the "target"}

\item{remove_no_targets}{Removes indicators without a "target" value specified, default: TRUE}
}
\description{
internal function for calculating the extra fields needed for bullet charts
}
\details{
internal function for calculating the extra fields for the bullet chart
}
